requirements.txt
requests==2.21.0
python-dotenv==0.10.3
pillow==6.2.0